enforce one scene action per match/unmatch of an entry; ui tweaks and warnings fixes

+++ src/pm-match-reaction-widget.hpp
@@ -114,7 +114,7 @@ protected:
     PmReactionType m_reactionType;
     size_t m_matchIndex = (size_t)-1;
     size_t m_multiConfigSz = 0;
-    int m_lastActionCount = -1;
+    size_t m_lastActionCount = (size_t)-1;
 
     PmCore *m_core;
     PmAddActionMenu *m_addActionMenu;
